refactor: 도메인 관련 내부 구조 변경

+++ src/main/java/com/pdev/atoz/order/dto/ProductCreateDto.java
@@ -5,7 +5,7 @@ import com.pdev.atoz.order.domain.Category;
 public record ProductCreateDto(
         String productName,
         Category category,
-        long price,
+        int price,
         String description
 ) {
 }
